+++ src/core/endpoints/file_upload/download_file.ts
@@ -100,6 +100,8 @@ export class DownloadFileRequest<
       name,
     } = this.parameters;
 
-    return `/v1/files/${subscribeKey}/channels/${encodeString(channel)}/files/${id}/${name}`;
+    return `/v1/files/${subscribeKey}/channels/${encodeString(channel)}/files/${encodeString(id)}/${encodeString(
+      name,
+    )}`;
   }
 }