+++ src/crypto/modules/WebCryptoModule/aesCbcCryptor.ts
@@ -166,6 +166,6 @@ export default class AesCbcCryptor implements ICryptor {
    * @returns Serialized cryptor information.
    */
   toString() {
-    return `AesCbcCryptor { cipherKey: ${this.cipherKey} }`;
+    return 'AesCbcCryptor {}';
   }
 }